+++ src/Ctoken/JumpRateModel.cs
@@ -1,6 +1,7 @@
 using Neo;
 using Neo.SmartContract.Framework;
 using Neo.SmartContract.Framework.Services;
+using System.Numerics;
 
 namespace Ctoken
 {
@@ -14,7 +15,7 @@ namespace Ctoken
 
 
 
-        public static void PutInterestAttribute(uint _multiplierPerBlock, uint _baseRatePerBlock, uint _jumpMultiplierPerBlock, uint _kink)
+        public static void PutInterestAttribute(BigInteger _multiplierPerBlock, BigInteger _baseRatePerBlock, BigInteger _jumpMultiplierPerBlock, BigInteger _kink)
         {
             UInt160 InterestModelAddress = InterestModel.Get();
             Contract.Call(InterestModelAddress, "PutInterestAttribute", CallFlags.All, new object[] { _multiplierPerBlock, _baseRatePerBlock, _jumpMultiplierPerBlock , _kink });
@@ -36,7 +37,7 @@ namespace Ctoken
         /// <param name="borrows">The amount of borrows in the market</param>
         /// <param name="reserves">The amount of reserves in the market (currently unused)</param>
         /// <returns>The utilization rate as a mantissa between [0, 1e8]</returns>
-        public static ulong utilizationRate(ulong cash, ulong borrows, ulong reserves)
+        public static BigInteger utilizationRate(BigInteger cash, BigInteger borrows, BigInteger reserves)
         {
             UInt160 InterestModelAddress = InterestModel.Get();
             object utilizationRateObj = Contract.Call(InterestModelAddress, "getInterestAttribute", CallFlags.All, new object[] { cash,borrows, reserves });
@@ -51,11 +52,11 @@ namespace Ctoken
         /// <param name="borrows">The amount of borrows in the market</param>
         /// <param name="reserves">The amount of reserves in the market (currently unused)</param>
         /// <returns>The borrow rate percentage per block as a mantissa (scaled by 1e8)</returns>
-        public static ulong getBorrowRate(ulong cash, ulong borrows, ulong reserves)
+        public static BigInteger getBorrowRate(BigInteger cash, BigInteger borrows, BigInteger reserves)
         {
             UInt160 InterestModelAddress = InterestModel.Get();
             object BorrowRateObj = Contract.Call(InterestModelAddress, "getBorrowRate", CallFlags.All, new object[] { cash, borrows, reserves });
-            return (ulong)BorrowRateObj;
+            return (BigInteger)BorrowRateObj;
         }
 
         /// <summary>
@@ -66,11 +67,11 @@ namespace Ctoken
         /// <param name="reserves">The amount of reserves in the market</param>
         /// <param name="reserveFactorMantissa">The current reserve factor for the market</param>
         /// <returns>The supply rate percentage per block as a mantissa (scaled by 1e8)</returns>
-        public static ulong getSupplyRate(ulong cash, ulong borrows, ulong reserves, ulong reserveFactorMantissa)
+        public static BigInteger getSupplyRate(BigInteger cash, BigInteger borrows, BigInteger reserves, BigInteger reserveFactorMantissa)
         {
             UInt160 InterestModelAddress = InterestModel.Get();
             object SupplyRateObj = Contract.Call(InterestModelAddress, "getSupplyRate", CallFlags.All, new object[] { cash, borrows, reserves, reserveFactorMantissa });
-            return (ulong)SupplyRateObj;
+            return (BigInteger)SupplyRateObj;
         }
     }
 }
